+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [0.0.2] - 11/03/2021
 
 * Resolvers now allowed on specific routes
-* Breaking: Child router now uses the normal constructor XRouter
+* Add addChildren method on the XRouter
 
 ## [0.0.1] - 05/03/2021